Rearrenge MainWindowHeader & MainViewExtensionNotice custom controls
