--- RetroLinker/Styles/MainWindowHeader.axaml
@@ -1,48 +1,0 @@
-<Styles xmlns="https://github.com/avaloniaui"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:controls="using:RetroLinker.Styles"
-        x:DataType="controls:MainWindowHeader"
-        x:CompileBindings="True">
-    <Design.PreviewWith>
-        <Border Padding="0">
-            <controls:MainWindowHeader Width="300" Title="Testing" Classes="Header">
-                <controls:MainWindowHeader.RightControl>
-                    <Label Content="aaa"/>
-                </controls:MainWindowHeader.RightControl>
-                <controls:MainWindowHeader.LeftControl>
-                    <Label Content="bbb"/>
-                </controls:MainWindowHeader.LeftControl>
-            </controls:MainWindowHeader>
-        </Border>
-    </Design.PreviewWith>
-
-    <Style Selector="controls|MainWindowHeader">
-        <!-- Set Defaults -->
-        <Setter Property="Template">
-            <ControlTemplate>
-                <Grid Height="45" ColumnDefinitions="Auto,*,Auto">
-                    <StackPanel Grid.Column="1" Orientation="Horizontal" 
-                                VerticalAlignment="Center" HorizontalAlignment="Center">
-                        <Label x:Name="HeaderLabel" Content="{TemplateBinding Title}"/>
-                    </StackPanel>
-                    
-                    <ContentControl x:Name="LeftControl" Grid.Column="0" Content="{TemplateBinding LeftControl}" 
-                                    VerticalAlignment="Center" HorizontalAlignment="Center"/>
-                    <ContentControl x:Name="RightControl" Grid.Column="2" Content="{TemplateBinding RightControl}"
-                                VerticalAlignment="Center" HorizontalAlignment="Center"/>
-                </Grid>
-            </ControlTemplate>
-        </Setter>
-    </Style>
-    
-    <Style Selector="controls|MainWindowHeader.Header">
-        <Style Selector="^ Grid">
-            <Setter Property="Background" Value="Firebrick"/>
-        
-            <Style Selector="^ Label#HeaderLabel">
-                <Setter Property="FontSize" Value="16"/>
-                <Setter Property="FontWeight" Value="Bold"/>
-            </Style>
-        </Style>
-    </Style>
-</Styles>
